Add alternative complexity head.
The current head is very wasteful, as the state is directly connected to an FC layer with no downsampling. This can result in a very top heavy network for small to medium sized networks. For example, in a 32x4x128 network the head account for 75% of the parameters, and the checkpoint size is 4MB. With this alternative head the checkpoint size is down to 1MB.